Add --max-iterations-per-second (#132)
<!--- Note to EXTERNAL Contributors -->
<!-- Thanks for opening a PR! 
If it is a significant code change, please **make sure there is an open
issue** for this.
We work best with you when we have accepted the idea first before you
code. -->

<!--- For ALL Contributors 👇 -->

## What was changed
<!-- Describe what has changed in this PR -->

Add RPS parameter for executor.

## Why?
<!-- Tell your future self why have you made these changes -->

To generate continuous but throttled load.

The new parameter allows to create a scenario where a large number of
iterations are started one-by-one at a constant rate (e.g. 5/s).

This is useful for creating a backlog of workflows (by turning off the
workers but keep creating new workflows).

## Checklist
<!--- add/delete as needed --->

1. Closes <!-- add issue number here -->

2. How was this tested:
<!--- Please describe how you tested your changes/how we can test them
-->

3. Any docs updates needed?
<!--- update README if applicable
      or point out where to update docs.temporal.io -->