Make throughputStress resumable (#143)
<!--- Note to EXTERNAL Contributors -->
<!-- Thanks for opening a PR! 
If it is a significant code change, please **make sure there is an open
issue** for this.
We work best with you when we have accepted the idea first before you
code. -->

<!--- For ALL Contributors 👇 -->

## What was changed
<!-- Describe what has changed in this PR -->

Added affordances to make the throughputStress scenario resumable.

## Why?
<!-- Tell your future self why have you made these changes -->

If the throughputStress scenario crashes (e.g. pod restarted), we want
to be able to continue where it left off.

## Test

Started Omes with

```
go run ./cmd run-scenario-with-worker \
--scenario throughput_stress \
--run-id omes \
--language go \
--iterations 2 \
--max-concurrent 1 \
--option internal-iteratios-timeout=1s \
--option visibility-count-timeout=5s
```

Once the first iteration succeeded, I stopped it and then ran (numbers
taken from the logged status update):

```
go run ./cmd run-scenario-with-worker \
--scenario throughput_stress \
--run-id omes \
--language go \
--iterations 2 \
--option internal-iteratios-timeout=1s \
--option visibility-count-timeout=5s \
--option internal-iterations-resume-from=1 \
--option workflow-count-resume-from=15
```

And it completed successfully.