Make throughputStress resumable (part 2) (#144)
<!--- Note to EXTERNAL Contributors -->
<!-- Thanks for opening a PR! 
If it is a significant code change, please **make sure there is an open
issue** for this.
We work best with you when we have accepted the idea first before you
code. -->

<!--- For ALL Contributors 👇 -->

## What was changed
<!-- Describe what has changed in this PR -->

Some tweaks to the previous PR
https://github.com/temporalio/omes/pull/143

I decided to make the scenario take care of tracking the state exactly
after all, as it allows the caller to be oblivious to its inner details.

## Checklist
<!--- add/delete as needed --->

Using it like this

<img width="802" alt="Screenshot 2025-04-30 at 6 33 05 PM"
src="https://github.com/user-attachments/assets/70465823-0a50-4962-8811-64de6d98ebd0"
/>