Set StartToCloseTimeout in ThroughputStressWorkflow (#150)
<!--- Note to EXTERNAL Contributors -->
<!-- Thanks for opening a PR! 
If it is a significant code change, please **make sure there is an open
issue** for this.
We work best with you when we have accepted the idea first before you
code. -->

<!--- For ALL Contributors 👇 -->

## What was changed
<!-- Describe what has changed in this PR -->

Set activity `StartToCloseTimeout` based on Workflow timeout.

## Why?
<!-- Tell your future self why have you made these changes -->

I [had removed the
StartToCloseTimeout](https://github.com/temporalio/omes/pull/146) but
now learned that it's required ...

```
BadScheduleActivityAttributes: A valid StartToClose or ScheduleToCloseTimeout is not set on ScheduleActivityTaskCommand.
```

This time around I ran it locally and confirmed this works.

## Checklist
<!--- add/delete as needed --->

1. Closes <!-- add issue number here -->

2. How was this tested:
<!--- Please describe how you tested your changes/how we can test them
-->

3. Any docs updates needed?
<!--- update README if applicable
      or point out where to update docs.temporal.io -->