Tweak ThroughputStress workflow (#156)
<!--- Note to EXTERNAL Contributors -->
<!-- Thanks for opening a PR! 
If it is a significant code change, please **make sure there is an open
issue** for this.
We work best with you when we have accepted the idea first before you
code. -->

<!--- For ALL Contributors 👇 -->

## What was changed
<!-- Describe what has changed in this PR -->

Tweaked the ThroughputStress workflow.

_PS: Note that I can't test PRs of Omes Workflow changes in our CICD
pipeline since that requires Docker images; which are not built for
PRs._

## Why?
<!-- Tell your future self why have you made these changes -->

(1) Improved Signal handling to address warning logs about unhandled
signals.
(2) Improved handling of Continue-as-New corner case to prevent error.

## Checklist
<!--- add/delete as needed --->

1. Closes <!-- add issue number here -->

2. How was this tested:
<!--- Please describe how you tested your changes/how we can test them
-->

3. Any docs updates needed?
<!--- update README if applicable
      or point out where to update docs.temporal.io -->