Fix SleepActivityConfig (#166)
<!--- Note to EXTERNAL Contributors -->
<!-- Thanks for opening a PR! 
If it is a significant code change, please **make sure there is an open
issue** for this.
We work best with you when we have accepted the idea first before you
code. -->

<!--- For ALL Contributors 👇 -->

## What was changed
<!-- Describe what has changed in this PR -->

Using a random seed instead of wrapping randomly generated output in
`workflow.SideEffect`.

## Why?
<!-- Tell your future self why have you made these changes -->

`workflow.SideEffect` doesn't work with arrays of proto messages. The
serialization breaks the result in this case, as one of the fields is
not JSON exportable.

## Checklist
<!--- add/delete as needed --->

**Reviewer: Looks bigger than it really is. I marked the two actual
changes.**