Check for negative duration (#169)
<!--- Note to EXTERNAL Contributors -->
<!-- Thanks for opening a PR! 
If it is a significant code change, please **make sure there is an open
issue** for this.
We work best with you when we have accepted the idea first before you
code. -->

<!--- For ALL Contributors 👇 -->

## What was changed
<!-- Describe what has changed in this PR -->

Check that `Duration` is never negative.

## Why?
<!-- Tell your future self why have you made these changes -->

There's a subtle bug in `GenericExecutor`: when passing in a negative
Duration, it won't set up the context as it's assuming `Iterations` is
set - and start executing.

We could adjust `GenericExecutor`, but it's better to catch the issue
early.

## Checklist
<!--- add/delete as needed --->

1. Closes <!-- add issue number here -->

2. How was this tested:
<!--- Please describe how you tested your changes/how we can test them
-->

3. Any docs updates needed?
<!--- update README if applicable
      or point out where to update docs.temporal.io -->