throughput_stress retries (#168)
<!--- Note to EXTERNAL Contributors -->
<!-- Thanks for opening a PR! 
If it is a significant code change, please **make sure there is an open
issue** for this.
We work best with you when we have accepted the idea first before you
code. -->

<!--- For ALL Contributors 👇 -->

## What was changed
<!-- Describe what has changed in this PR -->

Added retries for the Workflow Start.

## Why?
<!-- Tell your future self why have you made these changes -->

When we run the throughput_stress scenario with faults in our CICD
pipeline, the workflow start can fail. Omes will abort the run
immediately. CICD will restart Omes and resume.

All that creates unnecessary delays and noise. Internal retries mitigate
that.

## Checklist
<!--- add/delete as needed --->

1. Closes <!-- add issue number here -->

2. How was this tested:
<!--- Please describe how you tested your changes/how we can test them
-->

3. Any docs updates needed?
<!--- update README if applicable
      or point out where to update docs.temporal.io -->