throughput_stress skip run if out of time (#170)
<!--- Note to EXTERNAL Contributors -->
<!-- Thanks for opening a PR! 
If it is a significant code change, please **make sure there is an open
issue** for this.
We work best with you when we have accepted the idea first before you
code. -->

<!--- For ALL Contributors 👇 -->

## What was changed
<!-- Describe what has changed in this PR -->

When resuming, skip executor when out of time.

## Why?
<!-- Tell your future self why have you made these changes -->

When resuming, it can happen that a failure occurs in-between the last
scenario run and the final post-scenario verification; and that the
duration for the run has elapsed.

In that case, we don't want to start anymore iterations; and skip
straight to the post-scenario verification.

But if we're unlucky, our last state could be stale and we expect fewer
workflows than there actually are; thus, never completing successfully.
In that case, we want to expect a _minimum_ count of workflows.

## Checklist
<!--- add/delete as needed --->

1. Closes <!-- add issue number here -->

2. How was this tested:
<!--- Please describe how you tested your changes/how we can test them
-->

3. Any docs updates needed?
<!--- update README if applicable
      or point out where to update docs.temporal.io -->