Verify fairness (#174)
<!--- Note to EXTERNAL Contributors -->
<!-- Thanks for opening a PR! 
If it is a significant code change, please **make sure there is an open
issue** for this.
We work best with you when we have accepted the idea first before you
code. -->

<!--- For ALL Contributors 👇 -->

## What was changed
<!-- Describe what has changed in this PR -->

Changes the ebb and flow scenario to:

- verify workflow counts after the scenario finished
- track schedule-to-start latency
- report and evaluate fairness indicators 

## Why?
<!-- Tell your future self why have you made these changes -->

To monitor and verify scenario success.

Note that the fairness verification will not fail the scenario run; it
is purely informational (at this point). Which fairness indicator is the
most useful isn't clear yet; we'll have to monitor and come back here to
adjust things.

## Checklist
<!--- add/delete as needed --->

1. Closes <!-- add issue number here -->

2. How was this tested:
<!--- Please describe how you tested your changes/how we can test them
-->

3. Any docs updates needed?
<!--- update README if applicable
      or point out where to update docs.temporal.io -->