Scenario validation (#185)
<!--- Note to EXTERNAL Contributors -->
<!-- Thanks for opening a PR! 
If it is a significant code change, please **make sure there is an open
issue** for this.
We work best with you when we have accepted the idea first before you
code. -->

<!--- For ALL Contributors 👇 -->

## What was changed
<!-- Describe what has changed in this PR -->

Added new `Configurable` interface for a scenario to opt-in to. It
exposes a method parse the scenario input.

## Why?
<!-- Tell your future self why have you made these changes -->

In a CICD context it's helpful to validate if a scenario is correctly
configured before starting it since any issues would only be surfaced
once the infra is all setup (time consuming and expensive).

## Checklist
<!--- add/delete as needed --->

1. Closes <!-- add issue number here -->

2. How was this tested:
<!--- Please describe how you tested your changes/how we can test them
-->

3. Any docs updates needed?
<!--- update README if applicable
      or point out where to update docs.temporal.io -->