Kitchensink test (#180)
<!--- Note to EXTERNAL Contributors -->
<!-- Thanks for opening a PR! 
If it is a significant code change, please **make sure there is an open
issue** for this.
We work best with you when we have accepted the idea first before you
code. -->

<!--- For ALL Contributors 👇 -->

## What was changed
<!-- Describe what has changed in this PR -->

Framework for checking kitchensink actions across SDKs.

(more tests to be added later ...)

## Why?
<!-- Tell your future self why have you made these changes -->

To ensure actions are supported correctly and comprehensively; and
document missing implementations.

Especially for new actions this helps ensure you've implemented them
correctly.

## Checklist
<!--- add/delete as needed --->

1. Closes <!-- add issue number here -->

2. How was this tested:
<!--- Please describe how you tested your changes/how we can test them
-->

3. Any docs updates needed?
<!--- update README if applicable
      or point out where to update docs.temporal.io -->