Fix flaky test (#190)
<!--- Note to EXTERNAL Contributors -->
<!-- Thanks for opening a PR! 
If it is a significant code change, please **make sure there is an open
issue** for this.
We work best with you when we have accepted the idea first before you
code. -->

<!--- For ALL Contributors 👇 -->

## What was changed
<!-- Describe what has changed in this PR -->

Bumped to Go 1.25 to use the new `synctest.Test` primitive to run
deterministic timing tests.

Tip: use `Hide Whitespace` option for reviewing.

## Why?
<!-- Tell your future self why have you made these changes -->

`TestRunDurationWithoutTimeout` is flaky - example:

https://github.com/temporalio/omes/actions/runs/17052148233/job/48342124492?pr=179#step:4:125

## Checklist
<!--- add/delete as needed --->

1. Closes <!-- add issue number here -->

2. How was this tested:
<!--- Please describe how you tested your changes/how we can test them
-->

3. Any docs updates needed?
<!--- update README if applicable
      or point out where to update docs.temporal.io -->