Port throughput_stress to kitchenSink (#171)
<!--- Note to EXTERNAL Contributors -->
<!-- Thanks for opening a PR! 
If it is a significant code change, please **make sure there is an open
issue** for this.
We work best with you when we have accepted the idea first before you
code. -->

<!--- For ALL Contributors 👇 -->

## What was changed
<!-- Describe what has changed in this PR -->

Ported the throughput_stress scenario to use kitchensink.

## Why?
<!-- Tell your future self why have you made these changes -->

(1) Allows re-using it in other SDKs.
(2) Moves all the scenario code into one place.
(3) Makes scenario generally easier to test in CICD since a worker
change might not be needed.
(4) Improves debugging since inputs are instructive and readable (at
least most of them).

## Checklist
<!--- add/delete as needed --->

1. Closes <!-- add issue number here -->

2. How was this tested:
<!--- Please describe how you tested your changes/how we can test them
-->

Ran a longer local test, too:

```
go run ./cmd run-scenario-with-worker \
  --scenario throughput_stress \
  --language go \
  --iterations 50 \
  --max-concurrent 20 \
  --option internal-iterations=5 \
  --option continue-as-new-after-iterations=2 \
  --option sleep-time=1ms
```

Some of the kitchensink tests seem to be flaky, though ... will need to
address that.

3. Any docs updates needed?
<!--- update README if applicable
      or point out where to update docs.temporal.io -->