Remove scenario name from task queue name (#197)
<!--- Note to EXTERNAL Contributors -->
<!-- Thanks for opening a PR! 
If it is a significant code change, please **make sure there is an open
issue** for this.
We work best with you when we have accepted the idea first before you
code. -->

<!--- For ALL Contributors 👇 -->

## What was changed
<!-- Describe what has changed in this PR -->

Omes client and worker get their task queue _only_ from the RunID now;
instead of also the scenario name.

## Why?
<!-- Tell your future self why have you made these changes -->

RunID is already expected to be a unique identifier; scenario name is
superfluous. When running multiple scenarios, the user can pick a
unique, clear RunID to distinguish them.

This also simplifies the CICD use case where the workers need to be
deployed first before the clients are run. That means the name(s) of the
scenario(s) need to known and propagated ahead of time. Now all that's
needed is a RunID.

## Checklist
<!--- add/delete as needed --->

1. Closes <!-- add issue number here -->

2. How was this tested:
<!--- Please describe how you tested your changes/how we can test them
-->

3. Any docs updates needed?
<!--- update README if applicable
      or point out where to update docs.temporal.io -->