Reusable CLI (#214)
<!--- Note to EXTERNAL Contributors -->
<!-- Thanks for opening a PR! 
If it is a significant code change, please **make sure there is an open
issue** for this.
We work best with you when we have accepted the idea first before you
code. -->

<!--- For ALL Contributors 👇 -->

## What was changed
<!-- Describe what has changed in this PR -->

Moved CLI commands to exported package:

- new package `cli`
- renamed `cmdoptions` to `clioptions` to match

## Why?
<!-- Tell your future self why have you made these changes -->

This allows writing a thin Omes CLI wrapper application that defines its
own scenario; but still provide all the Omes CLI options as if it was an
"internal" scenario.

Of course this only works if the scenario uses the kitchensink executor.

```go
package main
 
import (
    "github.com/temporalio/omes/cmd/cli"
     _ "examples/custom-scenario/scenarios" // register our custom scenarios
)

func main() {
    cli.Main()
}
```

Using Omes programmatically as a library was already possible; but the
CLI use case was not.

## Checklist
<!--- add/delete as needed --->

1. Closes <!-- add issue number here -->

2. How was this tested:
<!--- Please describe how you tested your changes/how we can test them
-->

3. Any docs updates needed?
<!--- update README if applicable
      or point out where to update docs.temporal.io -->