Add option to ignore already started error (#227)
<!--- Note to EXTERNAL Contributors -->
<!-- Thanks for opening a PR! 
If it is a significant code change, please **make sure there is an open
issue** for this.
We work best with you when we have accepted the idea first before you
code. -->

<!--- For ALL Contributors 👇 -->

## What was changed
<!-- Describe what has changed in this PR -->

WISOTT

## Why?
<!-- Tell your future self why have you made these changes -->

When running a scenario it can happen that transient failures lead to an
incomplete scenario run and on retry the scenario will then fail because
workflows were already created. It's not always possible here to use a
new/clean task queue; in those cases, this new flag allows to skip over
these errors.

## Checklist
<!--- add/delete as needed --->

1. Closes <!-- add issue number here -->

2. How was this tested:
<!--- Please describe how you tested your changes/how we can test them
-->

3. Any docs updates needed?
<!--- update README if applicable
      or point out where to update docs.temporal.io -->