Small fixes for throughput stress with python worker (#233)
## What was changed
Two small fixes to allow `throughput_stress` to run on the Python
worker:
- Provide a default `BackoffCoefficient` for the `DefaultLocalActivity`
action. The Go SDK provides default values for local activity retry
policy fields (if missing), this is not the case in other SDKs. As such,
we provide a default backoff coefficient when we create the action
- Pass search attributes through to child workflows on the python
worker. Child workflows on the python worker were not including the
throughput stress search attribute, causing the visibility count check
to fail (and consequently, the scenario).

## Why?
We want the python worker to be able to run the throughput stress
scenario successfully

2. How was this tested:
Previously:
```
go run ./cmd run-scenario-with-worker \
    --scenario throughput_stress \
    --language python \
    --iterations 10 \
    --option internal-iterations=10
```
failed. With these fixes it succeeds.