Various fixes for TS worker's kitchen sink/throughput stress workflow (#235)
## What was changed
Various fixes for TS worker's kitchen sink/throughput stress workflow
- set correct activity timeout values (no 0 values)
- add search attributes to child workflows so visibility check passes
- fix pass through payload converter to simply pass through payloads
(was breaking on its `fromPayload` path because no encoding was set, and
it did not keep what encoding the payload was prior to encoding. But we
don't need to encode at all 😃)

## Why?
Allows usage of TS worker when running `throughput_stress`. To be used
for future smoke/stress testing.

2. How was this tested:
Able to run:
```
go run ./cmd run-scenario-with-worker \
    --scenario throughput_stress \
    --language typescript \
    --iterations 10 \
    --option internal-iterations=10
```
which failed before

3. Any docs updates needed?
No