Add `worker-err-on-unimplemented` option (#252)
## What was changed
Add `worker-err-on-unimplemented` option to explicitly fail on
unimplemented actions for kitchen sink workflows. This is currently
scoped to handling concurrent client actions which are not implemented
in all workers except Go.

Behavior has changed to, by default, skip/no-op unimplemented concurrent
client actions instead of failing. Setting this option to `true` enables
the previous behavior (failing on unimplemented concurrent client
actions).

Opted for a worker option because it's the simplest to implement. Other
options included:
- creating an option for kitchen sink workflow, which would control a
new flag in the `WorkflowInput` proto, used to tell the worker whether
they are to ignore concurrent client actions
- creating an option in the fuzzer scenario which would control whether
the fuzzer could generate `ClientActivity` with concurrent action sets

but they're both more complex. We can revisit this if necessary, but
this is functional and a small diff.

Additional small fix for TS worker when spawning child workflows (found
when running the fuzzer)

## Why?
The fuzzer scenario will always trigger this error otherwise, which
makes it difficult to find actual bugs with the fuzzer

Resolves: https://github.com/temporalio/omes/issues/223

2. How was this tested:
Ran
```
go run ./cmd run-scenario-with-worker \
      --scenario fuzzer \
      --language {language} \
      --iterations 1
```
with and without the `--worker-err-on-unimplemented` flag.