Wait for ExecuteChildWorkflowAction to complete by default (#257)
<!--- Note to EXTERNAL Contributors -->
<!-- Thanks for opening a PR! 
If it is a significant code change, please **make sure there is an open
issue** for this.
We work best with you when we have accepted the idea first before you
code. -->

<!--- For ALL Contributors 👇 -->

## What was changed
<!-- Describe what has changed in this PR -->

Make .NET, Python and TS kitchensink workflows wait for child workflow
execution completion.

Go and Java already do the right thing.

## Why?
<!-- Tell your future self why have you made these changes -->

The default wait `AwaitableChoice` policy is `waitFinish`, but these
impls only wait for the start, not completion.

## Checklist
<!--- add/delete as needed --->

1. Closes <!-- add issue number here -->

2. How was this tested:
<!--- Please describe how you tested your changes/how we can test them
-->

3. Any docs updates needed?
<!--- update README if applicable
      or point out where to update docs.temporal.io -->

Signed-off-by: Stephan Behnke <stephanos@users.noreply.github.com>