Extend metrics schema (#273)
## What was changed
Minor cleanup/code movement and adding additional fields to
`MetricLine`.

Added fields still need to be plumbed through (will be empty for now):
- `Environment`: environment that the scenario was run from 
- `BuildID`: build ID of the worker
- `Scenario`: scenario name
- `RunConfigProfile`: profile name of the run configuration
- `WorkerConfigProfile`: profile name of the worker configuration

where a `profile` maps to a configuration.

Want some feedback to establish the metrics schema. The goal is to have
a schema we anticipate to not have to change.

Plumbing to come in a subsequent PR (in progress, will link).

## Why?
Allows us to filter queries on more parameters that will be useful.
Establish the schema for the data team to create omni topics for.

2. How was this tested:
```
go run ./cmd run-scenario-with-worker \
    --scenario throughput_stress \
    --run-id test-$(date +%s) \
    --language go \
    --duration 10m \
    --prom-listen-address 127.0.0.1:9091 \
    --worker-prom-listen-address 127.0.0.1:9092 \
--prom-instance-addr 127.0.0.1:9090 \
--prom-instance-config \
--prom-snapshot \
--prom-export-worker-metrics go-10m.parquet \
--prom-export-worker-job omes-worker \
--prom-export-metrics-step 3
```